feat: add task 5-merge-styles and fix 4-copy-directory

+++ 04-copy-directory/index.js
@@ -4,20 +4,32 @@ const folder = path.join(__dirname, 'files');
 const folderCopy = path.join(__dirname, 'files-copy');
 
 function copyFolderData(currFolder, copyFolder) {
+  fs.rm(copyFolder, { recursive: true, force: true } , (err) => {
+    if (err) throw err;
     fs.mkdir(copyFolder, {recursive: true}, (err) => {
-        if (err) throw err;
-        fs.readdir(currFolder, (err, files) => {
+      if (err) throw err;
+      fs.readdir(currFolder, (err, files) => {
         if (err) throw err;
         files.forEach(function copyFiles(file){
-            const copyFile = path.join(copyFolder, file);
-            const currFile = path.join(currFolder, file);
-            fs.copyFile(currFile, copyFile, (err) => {
-                 if (err) throw err;
-                 console.log('copy')
-                })
-            })
-         })
-    })
+          const copyFile = path.join(copyFolder, file);
+          const currFile = path.join(currFolder, file);
+          fs.stat(currFile, (error, stats) => {
+            if (error) {
+              console.log(error);
+            }else if (stats.isFile()){
+              fs.copyFile(currFile, copyFile, (err) => {
+                if (err) throw err;
+                console.log('copy');
+              });
+            }else if(stats.isDirectory()){
+              copyFolderData(currFile, copyFile);
+            }
+          });
+        });
+      });
+    });
+  });
+  
 }
-copyFolderData(folder, folderCopy)
+copyFolderData(folder, folderCopy);
